Read the grid by folding the signal, not by searching for it
The search this replaces could only propose grids that land exactly on the
far edge of the image. That is an assumption, and it is sometimes false: of
the sheets with a grid in the book, three hold one the search could not name
at all, whatever it scored.

Folding needs no such rule. Take the difference between neighbouring
columns, ask of each whether it is loud enough to be a boundary, and fold
that answer onto a candidate pitch. The true pitch drops every boundary into
one slot. Share alone would not separate a pitch from twice a pitch, so the
share is weighed by the root of the turns it took: an answer from two turns
is a guess, and an answer from fifty is a measurement.

The gap now comes from the fold instead of a search. A gap makes two slots
tall, the tile ends against it and it ends against the next tile, and the
way round from one to the other is the gap itself.

An offset can now run backwards. The fold cannot tell a margin from a tile
that the edge cut, so the near columns of the image decide: flat colour is a
margin, anything else is a tile that starts before the edge.

On the sheets with a grid in the book this gets 11 of 16 right, counting the
tile, the gap and the offset. Every sheet of three whole packs agrees with
its neighbours.

+++ src/sheet.rs
@@ -703,7 +703,7 @@ impl Sheet {
         let read = side.tile.is_none().then(|| detect::grid(&img, tile));
         let tile = side.tile.map(Pair::xy).unwrap_or_else(|| read.map_or(tile, |(x, y)| [x.tile, y.tile]));
         let gap = side.gap.map(Pair::xy).unwrap_or_else(|| read.map_or([0, 0], |(x, y)| [x.gap, y.gap]));
-        let found = read.map_or([0, 0], |(x, y)| [x.offset as i32, y.offset as i32]);
+        let found = read.map_or([0, 0], |(x, y)| [x.offset, y.offset]);
         let offset = clamp_offset(side.offset.map(Pair::xy).unwrap_or(found), tile, gap);
         let prov = ProvMap::from_side(img.width(), img.height(), &side.provenance);
         let mut s = Self {
